+++ pages/seller/orders/[orderId].jsx
@@ -1,5 +1,7 @@
 import Page from '../../../components/Page';
 import { useRouter } from 'next/router';
+import orders from './orders';
+
 const OrderDetails = () => {
   const router = useRouter()
   const id = router.query.orderId
