+++ pages/seller/orders/orders.js
@@ -1,0 +1,31 @@
+const orders = [
+    {
+      id: '2342',
+      name: 'Jensen Huang',
+      email: 'jensen.huang@gmail.com',
+      total: '₹ 200',
+      status: 'done',
+      date: '29 Aug 2022',
+      action: { name: 'View' },
+    },
+    {
+      id: '2340',
+      name: 'Alex Grim',
+      email: 'alex.grim@gmail.com',
+      total: '₹ 240',
+      status: 'cancelled',
+      date: '29 Aug 2022',
+      action: { name: 'View' },
+    },
+    {
+      id: '1923',
+      name: 'Saud Kadiri',
+      email: 'saud.kadiri@gmail.com',
+      total: '₹ 240',
+      status: 'pending',
+      date: '20 Jan 2019',
+      action: { name: 'View' },
+    },
+  ];
+
+export default orders;
